Add files via upload
This project focuses on the analysis of a comprehensive movie dataset comprising 10,866 rows and 21 columns. To ensure the accuracy and reliability of the analysis, meticulous data cleaning was performed. Following this, quantitative techniques were applied to reveal hidden insights within the dataset. The findings are presented through a series of charts and maps, which have been thoroughly analyzed to provide a deep understanding of the data.